==> 2_ms2_annotation/0_fixing_GNPS_output.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Fixing-GNPS-annotation-output" data-toc-modified-id="Fixing-GNPS-annotation-output-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Fixing GNPS annotation output</a></span><ul class="toc-item"><li><span><a href="#Import-libraries" data-toc-modified-id="Import-libraries-1.1"><span class="toc-item-num">1.1&nbsp;&nbsp;</span>Import libraries</a></span></li><li><span><a href="#Glob-GNPS-output-file" data-toc-modified-id="Glob-GNPS-output-file-1.2"><span class="toc-item-num">1.2&nbsp;&nbsp;</span>Glob GNPS output file</a></span></li><li><span><a href="#Select-columns-of-interest" data-toc-modified-id="Select-columns-of-interest-1.3"><span class="toc-item-num">1.3&nbsp;&nbsp;</span>Select columns of interest</a></span></li><li><span><a href="#Translate-InChIKey-to-xref-in-KEGG,-BioCyc,-HMDB,-etc." data-toc-modified-id="Translate-InChIKey-to-xref-in-KEGG,-BioCyc,-HMDB,-etc.-1.4"><span class="toc-item-num">1.4&nbsp;&nbsp;</span>Translate InChIKey to xref in KEGG, BioCyc, HMDB, etc.</a></span></li><li><span><a href="#Use-InChIKey-to-extract-Smiles,-Molecular-Formulas-from-PubChem" data-toc-modified-id="Use-InChIKey-to-extract-Smiles,-Molecular-Formulas-from-PubChem-1.5"><span class="toc-item-num">1.5&nbsp;&nbsp;</span>Use InChIKey to extract Smiles, Molecular Formulas from PubChem</a></span></li><li><span><a href="#Save" data-toc-modified-id="Save-1.6"><span class="toc-item-num">1.6&nbsp;&nbsp;</span>Save</a></span></li></ul></li></ul></div>

# Fixing GNPS annotation output

## Import libraries

import pandas as pd
import numpy as np
import glob
import requests
import time
import json
import io

## Glob GNPS output file

gnps_file = glob.glob('gnps/ProteoSAFe-MOLECULAR-LIBRARYSEARCH-*/*.tsv')[0]
gnps_file

# import
df = pd.read_csv(gnps_file, sep = '\t')

# remove positive hits
df = df[df['IonMode']=='Negative'].reset_index()

print(df.shape)
df.head()

## Select columns of interest

These columns match the output by Sirius as well.

subset_df = df.copy()

subset_df = subset_df.loc[:, [
    'Adduct',
    'IonMode',
    'InChIKey',
    '#Scan#',
    'superclass',
    'class',
    'subclass',
]]

subset_df = subset_df[~subset_df['InChIKey'].isna()].reset_index(drop=True)
subset_df['#Scan#'] = "FT"+subset_df['#Scan#'].apply(lambda x: '{0:0>4}'.format(x))
subset_df.rename(columns={'#Scan#':'Features'}, inplace=True)
subset_df = subset_df.set_index('Features')
subset_df = subset_df.reset_index()
list_columns = list(subset_df.columns)
new_columns = ['MolecularFormula', 'ExactMass', 'InChI', 'CanonicalSmiles', 
               'PubChem_CID', 'KEGG', 'BioCyc', 'LIPIDMAPS', 'ChEBI', 'HMDB']

list_columns.extend(new_columns)
subset_df = subset_df.reindex(columns = list_columns, fill_value='')

print(subset_df.shape)
subset_df.head()

## Translate InChIKey to xref in KEGG, BioCyc, HMDB, etc.

I used the REST API by [Chemical Translation Service - CTS](https://cts.fiehnlab.ucdavis.edu/)

for index, row in subset_df.iterrows():
    #print(index, row['InChIKey'])
    
    query = row['InChIKey']
    
    ## translate inchikey using the Chemical Translation Service - CTS
    s = requests.Session()
    main_url = "https://cts.fiehnlab.ucdavis.edu/rest/"

    list_dbs = ['KEGG', 'BioCyc', 'LIPIDMAPS', 'ChEBI', 'Human Metabolome Database', "PubChem CID"]
    result_dict = {key:None for key in list_dbs}
    result_dict['InChIKey'] = query

    for db in list_dbs:

        r = s.get(main_url+'convert/InChIKey/'+db+'/'+query)
        r = json.loads(r.text)

        result = dict()

        if type(r) == dict: # error
            pass
        else:
            hit_dict = r[0]
            #print(hit_dict)
            if 'results' in hit_dict:
                if len(hit_dict['results']) == 1:
                    result_dict[db] = hit_dict['results'][0]
                elif len(hit_dict['results']) > 1:
                    result_dict[db] = ",".join(hit_dict['results'])
                else:
                    result_dict[db] = ""
            else:
                result_dict[db] = ""
    
    subset_df.loc[index,'KEGG'] = result_dict['KEGG']
    subset_df.loc[index,'BioCyc'] = result_dict['BioCyc']
    subset_df.loc[index,'LIPIDMAPS'] = result_dict['LIPIDMAPS']
    subset_df.loc[index,'ChEBI'] = result_dict['ChEBI']
    subset_df.loc[index,'HMDB'] = result_dict['Human Metabolome Database']
    subset_df.loc[index,'PubChem_CID'] = result_dict['PubChem CID']
    
print('Done')
subset_df.head()

## Use InChIKey to extract Smiles, Molecular Formulas from PubChem

GNPS does not provide these info in their output.

I used the Pug REST Service by PubChem.

for index, row in subset_df.iterrows():
    #print(index, row['InChIKey'])
    
    query = row['InChIKey']
    
    ## search pubchem for molecular formula, exact mass, smiles and inchi using inchikey
    ## 'MolecularFormula', 'ExactMass', 'InChI', 'CanonicalSmiles',
    properties = 'MolecularFormula,MolecularWeight,ExactMass,CanonicalSMILES,InChI'

    r = s.get('https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/inchikey/'+query+'/property/'+properties+'/CSV').content

    result_df = pd.read_csv(io.StringIO(r.decode('utf-8')))
    
    if 'Status: 404' in result_df.columns:
        pass
    
    elif not result_df.empty == True:
        
        mol_formula = ",".join(list(result_df['MolecularFormula'].unique()))
        subset_df.loc[index,'MolecularFormula'] = mol_formula
        
        ex_mass = ",".join([str(x) for x in result_df['ExactMass'].unique()])
        subset_df.loc[index,'ExactMass'] = ex_mass
        
        can_smiles = ",".join(list(result_df['CanonicalSMILES'].unique()))
        subset_df.loc[index,'CanonicalSmiles'] = can_smiles
        
        inchi = ",".join(list(result_df['InChI'].unique()))
        subset_df.loc[index,'InChI'] = inchi
    
    else:
        pass
    
    
    if index%5 == 0:
        time.sleep(3)

    #print(index, row['InChIKey'], 'Done searches')
    
print('Done')
subset_df.head()

## Save

subset_df.to_csv('summary_output_GNPS.csv', index=False)
