==> 2_ms2_annotation/2_merge_ms2_annotation.ipynb <==
# You're not done... Merge GNPS and SIRIUS Outputs

I am using the SIRIUS output that is not `exploded`. Exploded is just for MAGI.

## Let's gooooo

import pandas as pd
import numpy as np
import os
import re
from natsort import natsorted
from molmass import Formula


# -------------------------------------------------------------------------
def om_calculations(df, list_elements):
    """
    Calculations of O:C, H:C, NOSC, GFE, etc.
    """
    
    df[list_elements] = df[list_elements].replace(np.nan,0)
    df[list_elements] = df[list_elements].astype(float)
    
    df['OC'] = df['O']/df['C']
    df['HC'] = df['H']/df['C']
    df['NOSC'] = - ((4*df['C'] + df['H'] - 3*df['N'] - 2*df['O'] + 5*df['P'] - 2*df['S']) / (df['C'])) + 4
    df['GFE']= -(28.5 * df['NOSC']) + 60.3
    
    #Koch, B. P. and Dittmar, T.: From mass to structure: an aromaticity index for high-resolution mass data of natural organic matter, Rapid Commun. Mass Spectrom., 20(5), 926–932, doi:10.1002/rcm.2386, 2006.
    df['DBE'] = 1 + 0.5*(2* df['C'] - df['H'] + df['N'] + df['P'])
    df['DBE_O'] = (1 + 0.5*(2* df['C'] - df['H'] + df['N'] + df['P'])) - df['O']
    df['AI'] = (1 + df['C'] - df['O'] - df['S'] - ( (df['H']+df['P']+df['N'])*0.5) ) / (df['C'] - df['O'] - df['S'] - df['N'] - df['P'])
    df['AImod']  = (1 + df['C'] - (df['O']*0.5) - df['S'] - ( (df['H']+df['P']+df['N'])*0.5) ) / (df['C'] - (df['O']*0.5) - df['S'] - df['N'] - df['P'])
    df['DBE_AI'] = 1 + df['C'] - df['O'] - df['S'] - (0.5*(df['H'] + df['N'] + df['P']))
    
    indices = ['OC', 'HC', 'NOSC', 'GFE', 'DBE', 'DBE_O', 'AI', 'AImod', 'DBE_AI']
    
    df[indices] = df[indices].replace([-np.inf,+np.inf],[0,0])
    
    return df


# -------------------------------------------------------------------------

## Get GNPS

and fix a few column names so they concatenate correctly.

gnps = pd.read_csv('summary_output_GNPS.csv')

gnps = gnps.sort_values('Features').reset_index(drop=True)

features_gnps = list(gnps['Features'].values)

gnps = gnps.drop(['IonMode','LIPIDMAPS', 'InChIKey'], axis=1)

print(gnps.shape)

gnps_rename_dict = {
    'CanonicalSmiles':'Smiles'
}

gnps = gnps.rename(columns=gnps_rename_dict)

gnps.head(2)

## Get SIRIUS

Need some data wrangling here too...

sirius = pd.read_csv('summary_output_SIRIUS.csv')

sirius = sirius.sort_values('Features')

print(sirius.shape)

for ft in features_gnps:
    for index,row in sirius.iterrows():
        if row['Features'] == ft:
            sirius.drop(index, axis=0, inplace=True)
        else:
            continue

print(sirius.shape)

sirius = sirius.drop(['links','all classifications',
                      'YMDB', 'KNApSAcK','PlantCyc'], axis=1)

sirius_rename_dict = {
    'molecularFormula':'MolecularFormula',
    'adduct':'Adduct', 
    'smiles':'Smiles'
}

sirius = sirius.rename(columns=sirius_rename_dict)

sirius = sirius.reset_index(drop=True)

sirius.head(2)

## Concat GNPS and SIRIUS

ms2_concat = pd.concat([sirius, gnps], axis=0).reset_index(drop=True)

ms2_concat = ms2_concat.replace('', np.nan)

ms2_concat = ms2_concat.sort_values('Features').reset_index(drop=True)

ms2_concat.to_csv('summary_ms2_annotation.csv', index=False)

ms2_concat.head()

## From molecular formula, extract elements and calculate energetics like GFE, NOSC

and save!

df = ms2_concat.copy()

list_formulas = list(df['MolecularFormula'].unique())

unique_elements = set()

for formula in list_formulas:
    el_composition = [x[0] for x in Formula(formula).composition()]
    for el in el_composition:
        unique_elements.add(el)
        
for el in unique_elements:
    df[el] = df['MolecularFormula'].str.extract('('+el+'[\d]{0,3})')
    df[el] = df[el].replace(el,'1')
    df[el] = df[el].replace(el,"", regex=True)
    
list_columns = ['MolecularFormula']
list_columns.extend(list(unique_elements))

df = om_calculations(df, list(unique_elements))

df.to_csv('summary_ms2_annotation_energetics.csv', index=False)

df.head()
